+++ packages/eslint-config-vue/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @brucesong/eslint-config-vue
 
+## 1.0.11
+
+### Patch Changes
+
+- disable react-a11y
+
 ## 1.0.10
 
 ### Patch Changes